backend/requirements.txt
asgiref==3.4.1
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
distro==1.6.0
Django==3.2.9
django-extra-fields==3.0.2
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
jsonschema==3.2.0
MarkupSafe==2.0.1
oauthlib==3.1.1
paramiko==2.8.0
Pillow==8.4.0
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.4.0
pyrsistent==0.18.0
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
texttable==1.6.4
uritemplate==4.1.1
urllib3==1.26.7
websocket-client==0.59.0
